refactor(uri): move `uri.js` from `src/` to `src/util/`
유틸리티 함수이므로, util폴더로 이동함.
그에 따라, 의존하는 코드도 경로를 수정함.
